deprecated folder is created
