+++ deprecated/append.py
@@ -12,7 +12,7 @@ from importlib import import_module
 helao_root = os.path.dirname(os.path.realpath(__file__))
 sys.path.append(os.path.join(helao_root, 'config'))
 sys.path.append(helao_root)
-from helao import launcher
+from helao_launch import launcher
 confPrefix = sys.argv[1]
 appendPrefix = sys.argv[2]
