+++ deprecated/celery_conf.py
